moved listview to menu fragment for accurate reading. generateFeed still broken
