+++ app/src/main/java/com/example/musiclovers/MenuFragment.java
@@ -5,6 +5,8 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.ListView;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
@@ -17,6 +19,7 @@ import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QuerySnapshot;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class MenuFragment extends Fragment {
@@ -25,6 +28,10 @@ public class MenuFragment extends Fragment {
     private SpotifyPlayer mSpotifyPlayer;
     private String TAG = "MenuFragment";
 
+    // Access Cloud Firestore
+    FirebaseFirestore db = FirebaseFirestore.getInstance();
+    // Create ListView for posts
+    ListView listView;
 
     @Override
     public View onCreateView(
@@ -65,6 +72,12 @@ public class MenuFragment extends Fragment {
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
+        ArrayList<String> arrayList = new ArrayList<>();
+        listView = view.findViewById(R.id.list_view);
+        arrayList = generateFeed();
+        ArrayAdapter arrayAdapter = new ArrayAdapter(requireActivity(), android.R.layout.simple_list_item_1, arrayList);
+        listView.setAdapter(arrayAdapter);
+
         binding.buttonFirst.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -80,5 +93,49 @@ public class MenuFragment extends Fragment {
         binding = null;
     }
 
+    // faux generate feed just so that the feed can be populated with stuff until the real one is fixed
+    public ArrayList<String> generateFeed(){
+        ArrayList<String> arrayList = new ArrayList<>();
+        arrayList.add("hello");
+        arrayList.add("User: test \n\nthis song rocks \n\nLink:");
+        arrayList.add("User: test \n\ntest \n\nLink: test");
+        arrayList.add("User: test \n\nthis song bangs \n\nLink: osu.edu");
+        arrayList.add("User: test \n\nThis is a really good song\n\nLink: youtube.com");
+        arrayList.add("User: musiclover \n\nsong good \n\nLink: https://open.spotify.com/track/5UqCQaDshqbIk3pkhy4Pjg?si=febf62e52b43462a");
+        arrayList.add("User: musiclover \n\nthis bangs \n\nLink: https://open.spotify.com/track/5LxvwujISqiB8vpRYv887S?si=a1cbffe31e0b408d");
+        arrayList.add("User: test \n\ndds \n\nLink: d");
+        return arrayList;
+    }
+
+    // Method to retrieve all posts from Firestore database --BROKEN
+    public void ginerateFeed() {
+        ArrayList<String> feed = new ArrayList<>();
+        db.collection("Posts")
+                .get()
+                .addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
+                    @Override
+                    public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
+                        Log.d(TAG, "onSuccess: Retrieving posts");
+                        List<DocumentSnapshot> snapshotList = queryDocumentSnapshots.getDocuments();
+                        for (DocumentSnapshot snapshot : snapshotList) {
+                            String post = "";
+                            Log.d(TAG, "onSuccess: " + snapshot.getData().toString());
+                            // get username
+                            post = snapshot.get("userID").toString() + "\n";
+                            // get post contents
+                            post = snapshot.get("text").toString() + "\n";
+                            // get spotify link
+                            post = snapshot.get("link").toString();
+                            feed.add(post);
+                        }
+                    }
+                })
+                .addOnFailureListener(new OnFailureListener() {
+                    @Override
+                    public void onFailure(@NonNull Exception e) {
+                        Log.e(TAG, "onFailure: ", e);
+                    }
+                });
+    }
 
 }
