+++ app/src/main/res/layout/activity_main.xml
@@ -7,10 +7,7 @@
     android:orientation="vertical"
     tools:context=".MainActivity">
 
-    <ListView
-        android:id="@+id/list_view"
-        android:layout_width="match_parent"
-        android:layout_height="595dp" />
+
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
