+++ app/src/main/res/layout/fragment_menu.xml
@@ -8,12 +8,18 @@
     tools:context=".MenuFragment"
 
     >
+    <ListView
+        android:id="@+id/list_view"
+        android:layout_width="match_parent"
+        android:layout_height="595dp"
+        tools:ignore="MissingConstraints" />
+
 
     <Button
         android:id="@+id/button_first"
         android:layout_width="318dp"
         android:layout_height="57dp"
-        android:layout_marginBottom="76dp"
+        android:layout_marginBottom="50dp"
         android:text="@string/Post"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
